We removed what we had initially in the template.
Replaced that with header that has two links; Component 1 and Component
2.

The routerLink is a property binding directive that takes in the name
of the routes we have defined in our RouterConfig decorator below.

We also have the router-outlet tags where the routed views go.

Our RouterConfig decorator defines; path:this is the route that will
appear on the browser, name:this is the name that we use in referring
to the route and it is the one we are using in the routerLink,
component:which is the component we want loaded and useAsDefault in the
first route to show that we want this particular route to be loaded the
first time